jaiz-ussd/jaizussd/settings/dev.py
from .base import *
import django_heroku
from decouple import config
import dj_database_url
import environ

env = environ.Env()
environ.Env.read_env(os.path.join('.env'))

SECRET_KEY = env('SECRET_KEY')

DEBUG = True

ALLOWED_HOSTS = []

DATABASES = {
    'default': {
        'ENGINE': 'django.db.backends.postgresql_psycopg2',
        'NAME': env('DATABASE_NAME'),
        'USER': env('DATABASE_USER'),
        'PASSWORD': env('DATABASE_PASSWORD'),
        'HOST': env('DATABASE_HOST'),
        'PORT': env('DATABASE_PORT'),
    }
}

REST_FRAMEWORK = {
    'DEFAULT_AUTHENTICATION_CLASSES': [
        'rest_framework.authentication.SessionAuthentication',
        'rest_framework_simplejwt.authentication.JWTAuthentication',
    ],
    'DEFAULT_PERMISSION_CLASSES': [
        'rest_framework.permissions.IsAuthenticated',
    ],
}

CORS_ALLOWED_ORIGINS = [
    "http://localhost:8080",
    "http://localhost:80",
    "http://localhost",
]

CORS_ALLOW_ALL_ORIGINS = True
# CSRF_TRUSTED_ORIGINS = CORS_ALLOWED_ORIGINS


SIMPLE_JWT = {
    'ACCESS_TOKEN_LIFETIME': timedelta(days=30),
    'REFRESH_TOKEN_LIFETIME': timedelta(days=60),
    'UPDATE_LAST_LOGIN': True,
    'AUTH_HEADER_TYPES': ('Bearer', 'Token',),
}
# API SETTINGS
API_URL = env('API_URL')
API_USERNAME = env('API_USERNAME')
API_PASSWORD = env('API_PASSWORD')
API_CLIENT_IP = env('API_CLIENT_IP')
JAIZ_BANK_CODE = env('JAIZ_BANK_CODE')

PIN_TRIES = env('PIN_TRIES')
CHARGES = env('CHARGES')

# Activate Django-Heroku.
django_heroku.settings(locals())

