jaiz-ussd/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.10
cryptography==37.0.2
dj-database-url==0.5.0
Django==4.1
django-cors-headers==3.13.0
django-crontab==0.7.1
django-decouple==2.1
django-environ==0.9.0
django-filter==21.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
humanize==4.3.0
idna==3.3
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0
xmltodict==0.13.0
